svn merge -c 19673 https://source.sakaiproject.org/svn/content/trunk U    content-api/api/src/java/org/sakaiproject/content/api/ContentCollection.java U    content-tool/tool/src/java/org/sakaiproject/content/tool/ResourcesAction.java U    content-impl/impl/src/java/org/sakaiproject/content/impl/DbContentService.java U    content-impl/impl/src/java/org/sakaiproject/content/impl/BaseContentService.java /Develop/sakai_2-3-x/content$ svn log -r 19673 https://source.sakaiproject.org/svn/content/trunk ------------------------------------------------------------------------ r19673 | jimeng@umich.edu | 2006-12-18 14:26:26 -0500 (Mon, 18 Dec 2006) | 3 lines
SAK-7431
Added new way to count members of a collection.  Old way actually counted all entities at all levels within a hierarchy of folders.  That was replaced with a method that correctly counted just immediate children of a folder, but it loaded all members of a collection into memory to count them.  This new method arrives at a count through a database query that looks for members of the collection.

------------------------------------------------------------------------



git-svn-id: https://source.sakaiproject.org/svn/content/branches/sakai_2-3-x@19707 66ffb92e-73f9-0310-93c1-f5514f145a0a