SAK-13913 Added log message for errorMessages from FormattedText.  Here's a sample message:
The HTML tag '<script>' is not allowed in formatted text.
The HTML tag '</script>' is not allowed in formatted text.

As the javadocs for FormattedText.processFormattedText() says, that message could be displayed to the user.  But the user action succeeds (after escaping the text to render the script tags safe) and there is no way currently implemented to get a message to the user from the context in which the method is involved.  We'll write it to the logs, along with the id ot the resource/collection in which it occurred.


git-svn-id: https://source.sakaiproject.org/svn/content/trunk@52267 66ffb92e-73f9-0310-93c1-f5514f145a0a